# apps/backend/app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Awade API",
    "description": "AI-powered educator support platform for African teachers",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/contexts/": {
      "post": {
        "tags": [
          "contexts"
        ],
        "summary": "Create Context",
        "description": "Create a new context for a lesson plan.",
        "operationId": "create_context_api_contexts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contexts/lesson-plan/{lesson_plan_id}": {
      "get": {
        "tags": [
          "contexts"
        ],
        "summary": "Get Contexts By Lesson Plan",
        "description": "Get all contexts for a specific lesson plan.",
        "operationId": "get_contexts_by_lesson_plan_api_contexts_lesson_plan__lesson_plan_id__get",
        "parameters": [
          {
            "name": "lesson_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contexts/{context_id}": {
      "get": {
        "tags": [
          "contexts"
        ],
        "summary": "Get Context",
        "description": "Get a specific context by ID.",
        "operationId": "get_context_api_contexts__context_id__get",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "contexts"
        ],
        "summary": "Update Context",
        "description": "Update a specific context.",
        "operationId": "update_context_api_contexts__context_id__put",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contexts"
        ],
        "summary": "Delete Context",
        "description": "Delete a specific context.",
        "operationId": "delete_context_api_contexts__context_id__delete",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contexts/lesson-plan/{lesson_plan_id}/submit": {
      "post": {
        "tags": [
          "contexts"
        ],
        "summary": "Submit Context",
        "description": "Submit context for a lesson plan (frontend endpoint).",
        "operationId": "submit_context_api_contexts_lesson_plan__lesson_plan_id__submit_post",
        "parameters": [
          {
            "name": "lesson_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContextCreate": {
        "properties": {
          "context_text": {
            "type": "string",
            "title": "Context Text",
            "description": "The context text content"
          },
          "context_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Type",
            "description": "Type of context (e.g., cultural, resources, student_background)"
          },
          "lesson_plan_id": {
            "type": "integer",
            "title": "Lesson Plan Id",
            "description": "ID of the lesson plan this context belongs to"
          }
        },
        "type": "object",
        "required": [
          "context_text",
          "lesson_plan_id"
        ],
        "title": "ContextCreate",
        "description": "Schema for creating a new context."
      },
      "ContextListResponse": {
        "properties": {
          "contexts": {
            "items": {
              "$ref": "#/components/schemas/ContextResponse"
            },
            "type": "array",
            "title": "Contexts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "contexts",
          "total"
        ],
        "title": "ContextListResponse",
        "description": "Schema for list of contexts response."
      },
      "ContextResponse": {
        "properties": {
          "context_text": {
            "type": "string",
            "title": "Context Text",
            "description": "The context text content"
          },
          "context_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Type",
            "description": "Type of context (e.g., cultural, resources, student_background)"
          },
          "context_id": {
            "type": "integer",
            "title": "Context Id"
          },
          "lesson_plan_id": {
            "type": "integer",
            "title": "Lesson Plan Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "context_text",
          "context_id",
          "lesson_plan_id",
          "created_at",
          "updated_at"
        ],
        "title": "ContextResponse",
        "description": "Schema for context response."
      },
      "ContextSubmissionRequest": {
        "properties": {
          "context_text": {
            "type": "string",
            "title": "Context Text",
            "description": "The context text content"
          },
          "context_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Type",
            "description": "Type of context"
          }
        },
        "type": "object",
        "required": [
          "context_text"
        ],
        "title": "ContextSubmissionRequest",
        "description": "Schema for submitting context from frontend."
      },
      "ContextUpdate": {
        "properties": {
          "context_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Text",
            "description": "The context text content"
          },
          "context_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Type",
            "description": "Type of context"
          }
        },
        "type": "object",
        "title": "ContextUpdate",
        "description": "Schema for updating an existing context."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}